+++ envs/container/docker/av-rtmp-source/alpine/avtool.sh
@@ -7,7 +7,7 @@
 #
 # executable
 ###########################################################################################################################################################################################
-set -eu
+set -u
 repoRoot="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 cd "$repoRoot"
 #######################################################
@@ -64,19 +64,31 @@ checkError() {
         exit 1
     fi
 }
-
+checkDevContainerMode () {
+    checkDevContainerModeResult="0"
+    VOLNAME=$(docker volume inspect av-services_devcontainer_tempvol --format {{.Name}} 2> /dev/null) || true    
+    # if running in devcontainer connect container to dev container network av-services_devcontainer_default
+    if [[ $VOLNAME == 'av-services_devcontainer_tempvol' ]] ; then
+        checkDevContainerModeResult="1"
+    fi
+    return
+}
 AV_SERVICE=av-rtmp-source
 AV_FLAVOR=alpine
 AV_IMAGE_NAME=${AV_SERVICE}-${AV_FLAVOR} 
 AV_IMAGE_FOLDER=av-services
 AV_CONTAINER_NAME=av-rtmp-source-alpine-container
-AV_RTMP_URL=rtmp://localhost:1935/live/stream
+AV_RTMP_PORT=1935
+AV_RTMP_HOST=localhost
+AV_RTMP_URL=rtmp://${AV_RTMP_HOST}:${AV_RTMP_PORT}/live/stream
 # Check if configuration file exists
 if [[ ! -f "$repoRoot"/"$configuration_file" ]]; then
     cat > "$repoRoot"/"$configuration_file" << EOF
 AV_IMAGE_NAME=${AV_IMAGE_NAME}
 AV_IMAGE_FOLDER=${AV_IMAGE_FOLDER}
 AV_CONTAINER_NAME=${AV_CONTAINER_NAME}
+AV_RTMP_PORT=${AV_RTMP_PORT}
+AV_RTMP_HOST=${AV_RTMP_HOST}
 AV_RTMP_URL=${AV_RTMP_URL}
 AV_TEMPDIR=$(mktemp -d)
 EOF
@@ -85,6 +97,8 @@ fi
 export $(grep AV_IMAGE_NAME "$repoRoot"/"$configuration_file")
 export $(grep AV_IMAGE_FOLDER "$repoRoot"/"$configuration_file")
 export $(grep AV_CONTAINER_NAME "$repoRoot"/"$configuration_file")
+export $(grep AV_RTMP_PORT "$repoRoot"/"$configuration_file")
+export $(grep AV_RTMP_HOST "$repoRoot"/"$configuration_file")
 export $(grep AV_RTMP_URL "$repoRoot"/"$configuration_file")
 export $(grep AV_TEMPDIR "$repoRoot"/"$configuration_file" |  { read test; if [[ -z $test ]] ; then AV_TEMPDIR=$(mktemp -d) ; echo "AV_TEMPDIR=$AV_TEMPDIR" ; echo "AV_TEMPDIR=$AV_TEMPDIR" >> .avtoolconfig ; else echo $test; fi } )
 
@@ -95,38 +109,45 @@ fi
 
 if [[ "${action}" == "install" ]] ; then
     echo "Installing pre-requisite"
-    echo "Installing azure cli"
-    curl -sL https://aka.ms/InstallAzureCLIDeb | sudo bash
-    echo "Installing ffmpeg"
-    sudo apt-get -y update
-    sudo apt-get -y install ffmpeg
-    if [ ! -f "${AV_TEMPDIR}"/camera-300s.mkv ]; then
-        echo "Downloading content"
-        wget --quiet https://github.com/flecoqui/av-services/raw/main/content/camera-300s.mkv -O "${AV_TEMPDIR}"/camera-300s.mkv     
+    checkDevContainerMode  || true
+    if [[ "$checkDevContainerModeResult" == "1" ]] ; then
+        echo "As running in devcontainer av-services_devcontainer installation not required"
+        echo -e "${GREEN}Installing pre-requisites done${NC}"
+        exit 0
+    else        
+        echo "Installing azure cli"
+        curl -sL https://aka.ms/InstallAzureCLIDeb | sudo bash
+        echo "Installing ffmpeg"
+        sudo apt-get -y update
+        sudo apt-get -y install ffmpeg
+        if [ ! -f "${AV_TEMPDIR}"/camera-300s.mkv ]; then
+            echo "Downloading content"
+            wget --quiet https://github.com/flecoqui/av-services/raw/main/content/camera-300s.mkv -O "${AV_TEMPDIR}"/camera-300s.mkv     
+        fi
+        echo "Installing docker"
+        # removing old version
+        sudo apt-get remove docker docker-engine docker.io containerd runc
+        sudo apt-get -y update
+        sudo apt-get -y install \
+            apt-transport-https \
+            ca-certificates \
+            curl \
+            gnupg-agent \
+            software-properties-common
+        curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
+        sudo apt-key fingerprint 0EBFCD88
+        sudo add-apt-repository \
+            "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
+            $(lsb_release -cs) \
+            stable"
+        sudo apt-get -y update
+        sudo apt-get -y install  jq
+        sudo apt-get -y install docker.io    
+        sudo groupadd docker || true
+        sudo usermod -aG docker ${USER} || true
+        echo -e "${GREEN}Installing pre-requisites done${NC}"
+        exit 0
     fi
-    echo "Installing docker"
-    # removing old version
-    sudo apt-get remove docker docker-engine docker.io containerd runc
-    sudo apt-get -y update
-    sudo apt-get -y install \
-        apt-transport-https \
-        ca-certificates \
-        curl \
-        gnupg-agent \
-        software-properties-common
-    curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
-    sudo apt-key fingerprint 0EBFCD88
-    sudo add-apt-repository \
-        "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
-        $(lsb_release -cs) \
-        stable"
-    sudo apt-get -y update
-    sudo apt-get -y install  jq
-    sudo apt-get -y install docker.io    
-    sudo groupadd docker || true
-    sudo usermod -aG docker ${USER} || true
-    echo -e "${GREEN}Installing pre-requisites done${NC}"
-    exit 0
 fi
 if [[ "${action}" == "login" ]] ; then
     echo "Login..."
@@ -176,35 +197,58 @@ if [[ "${action}" == "stop" ]] ; then
 fi
 if [[ "${action}" == "test" ]] ; then
     rm -f "${AV_TEMPDIR}"/*.mp4
-    rm -f "${AV_TEMPDIR}"/*.mkv
     if [ ! -f "${AV_TEMPDIR}"/camera-300s.mkv ]; then
         echo "Downloading content"
         wget --quiet https://github.com/flecoqui/av-services/raw/main/content/camera-300s.mkv -O "${AV_TEMPDIR}"/camera-300s.mkv     
     fi
     echo "Starting RTMP sink ${AV_FLAVOR} container"
+
     sudo docker container start "av-rtmp-sink-${AV_FLAVOR}-container"  
     containerId=$(sudo docker ps -a --format "{{.ID}}" -f "name=av-rtmp-sink-${AV_FLAVOR}-container")
-    privateIpAddress=$(sudo docker container inspect $containerId  --format='{{range .NetworkSettings.Networks}}{{.IPAddress}}{{end}}')
-    echo "RTMP sink private IP address: ${privateIpAddress}"
+    checkDevContainerMode  || true
+    if [[ "$checkDevContainerModeResult" == "1" ]] ; then
+        sudo docker network connect av-services_devcontainer_default av-rtmp-sink-${AV_FLAVOR}-container 
+        CONTAINER_IP=$(sudo docker container inspect "$containerId" | jq -r '.[].NetworkSettings.Networks."av-services_devcontainer_default".IPAddress')
+    else
+        CONTAINER_IP=$(sudo docker container inspect "$containerId"  --format='{{range .NetworkSettings.Networks}}{{.IPAddress}}{{end}}')
+    fi  
+    echo "RTMP sink private IP address: ${CONTAINER_IP}"
     echo "Starting test-${AV_CONTAINER_NAME} container..."
     sudo  docker container stop "test-${AV_CONTAINER_NAME}" &> /dev/null || true
     sudo  docker container rm "test-${AV_CONTAINER_NAME}" &> /dev/null  || true
-    sudo docker run  -d -it -e RTMP_URL=rtmp://${privateIpAddress}:1935/live/stream  --name "test-${AV_CONTAINER_NAME}" ${AV_IMAGE_FOLDER}/${AV_IMAGE_NAME} 
+    sudo docker run  -d -it -e RTMP_URL=rtmp://${CONTAINER_IP}:${AV_RTMP_PORT}/live/stream  --name "test-${AV_CONTAINER_NAME}" ${AV_IMAGE_FOLDER}/${AV_IMAGE_NAME} 
+    if [[ "$checkDevContainerModeResult" == "1" ]] ; then
+        sudo docker network connect av-services_devcontainer_default "test-${AV_CONTAINER_NAME}"
+    fi  
 
     echo "Capture 20s of RTMP stream on the host machine..."
-    ffmpeg   -hide_banner -loglevel error -re -rw_timeout 20000000 -i ${AV_RTMP_URL}  -t 00:00:20  -c copy -flags +global_header -f segment -segment_time 10 -segment_format_options movflags=+faststart -reset_timestamps 1 "${AV_TEMPDIR}"/testrtmp%d.mp4 &
+    ffmpeg   -hide_banner -loglevel error -re -rw_timeout 20000000 -i rtmp://${CONTAINER_IP}:${AV_RTMP_PORT}/live/stream  -t 00:00:20  -c copy -flags +global_header -f segment -segment_time 10 -segment_format_options movflags=+faststart -reset_timestamps 1 "${AV_TEMPDIR}"/testrtmp%d.mp4 &
     sleep 40
     echo "Check mp4 captured streams in directory : ${AV_TEMPDIR}"
     if [[ ! -f "${AV_TEMPDIR}/testrtmp0.mp4" || ! -f "${AV_TEMPDIR}/testrtmp1.mp4" ]] ; then
+
+        # if running in devcontainer disconnect container from dev container network av-services_devcontainer_default
+        checkDevContainerMode  || true
+        if [[ "$checkDevContainerModeResult" == "1" ]] ; then
+            sudo docker network disconnect av-services_devcontainer_default av-rtmp-sink-${AV_FLAVOR}-container 
+            sudo docker network disconnect av-services_devcontainer_default "test-${AV_CONTAINER_NAME}"
+        fi    
         echo "Stopping RTMP sink ${AV_FLAVOR} container"
         sudo docker container "stop av-rtmp-sink-${AV_FLAVOR}-container"
         echo "Stopping test-${AV_CONTAINER_NAME} container..."
         sudo docker container stop "test-${AV_CONTAINER_NAME}"  
         echo "RTMP Test failed - check file ${AV_TEMPDIR}/testrtmp0.mp4"
+         
         kill %1
         exit 1
     fi
     echo "Stopping RTMP sink ${AV_FLAVOR} container"
+    # if running in devcontainer disconnect container from dev container network av-services_devcontainer_default
+    checkDevContainerMode  || true
+    if [[ "$checkDevContainerModeResult" == "1" ]] ; then
+        sudo docker network disconnect av-services_devcontainer_default av-rtmp-sink-${AV_FLAVOR}-container 
+        sudo docker network disconnect av-services_devcontainer_default "test-${AV_CONTAINER_NAME}"
+    fi       
     sudo docker container stop "av-rtmp-sink-${AV_FLAVOR}-container"  
     echo "Stopping test-${AV_CONTAINER_NAME} container..."
     sudo docker container stop "test-${AV_CONTAINER_NAME}"
